Now using neighbour struct
The neighbour struct is now in use. It has also been fixed to encapsulate precisely the information it is supposed to, the global variables associated with a neighbour.
Furthermore, a currently unused struct for encapsulating the neighbour data associated with the state of the selective repeat function has been created.
The variables that the neighbour struct replaces have been commented out and so have their associated code lines. The code lines that are commented out have had replacements using the structs instead installed.