Introduction & part of Design
The introduction should be good enough.
The design section contains the appropriate parts from project part 1 and 2, shortened down so they don't take as much space.
The are a few things, such as the signals used, that have been updated to include the part 3 information. The actual documentation of how part 3 has been built is missing though.